Improve float serialization.
Float serialization used the same serialization code as normal integers, which writes less bytes if the integer is of sufficiently small size. The problem however is that a float is basically *never* so small due to how the bits are laid out and this almost always resulted in a wasted extra byte vs just writing the 4/8 bytes of the float directly.

("basically never" is everything except 0/subnormals, which I'd say is not worth adding an extra byte to almost all other floats over. I even tested all possible floats just to make sure.)

The new code is also faster on Core because it writes the 4/8 bytes at once instead of individual WriteByte/ReadByte() calls (see benchmarks in PR).